requirements.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
daff==1.4.2
db-dtypes==1.5.0
fastjsonschema==2.21.2
google-api-core==2.30.0
google-auth==2.48.0
google-cloud==0.34.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.78.0
idna==3.11
leather==0.4.1
Logbook==1.5.3
numpy==2.4.2
oauthlib==3.3.1
ordered-set==4.1.0
packaging==26.0
pandas==2.3.3
parsedatetime==2.6
pathspec==0.12.1
platformdirs==4.9.2
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
setuptools==82.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
text-unidecode==1.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==16.0
zipp==3.23.0
